Moving extension generators into top level folders for better structure.
