+++ Remotr.SourceGen/StatelessExtensionGenerator/StatelessQueryExtensionGenerator.cs
@@ -3,7 +3,7 @@ using Microsoft.CodeAnalysis.CSharp.Syntax;
 using System.Text;
 using Remotr.SourceGen.Shared;
 
-namespace Remotr.SourceGen.UseHandlerAttributes.ExtensionGenerators;
+namespace Remotr.StatelessExtensionGenerators;
 
 /// <summary>
 /// Generator for StatelessQueryHandler types that generates appropriate extension methods.
